pyproject.toml
[project]
name = "docx-mcp-server"
version = "0.1.0"
description = "A Model Context Protocol server for editing Docx files"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=0.1.0",
    "python-docx>=1.1.0"
]

[project.scripts]
mcp-server-docx = "docx_mcp_server.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/docx_mcp_server"]
